requirements.txt
numpy
opencv-python
flask
pillow
flask-cors
torch
transformers
datasets
pandas